feat: sanitize server span uri names
